+++ main.js
@@ -7,7 +7,7 @@ import whatanime from './src/whatanime';
 import ascii2d from './src/ascii2d';
 import CQ from './src/CQcode';
 import psCache from './src/cache';
-import Logger from './src/Logger';
+import logger from './src/logger';
 import RandomSeed from 'random-seed';
 import sendSetu from './src/plugin/setu';
 import Akhr from './src/plugin/akhr';
@@ -25,7 +25,6 @@ import asyncMap from './src/utils/asyncMap';
 const ocr = require('./src/plugin/ocr');
 
 const bot = new CQWebSocket(global.config.cqws);
-const logger = new Logger();
 const rand = RandomSeed.create();
 
 // 全局变量
@@ -144,7 +143,7 @@ async function commonHandle(e, context) {
   if (context.user_id === bot._qq) return true;
 
   // 黑名单检测
-  if (Logger.checkBan(context.user_id, context.group_id)) return true;
+  if (logger.checkBan(context.user_id, context.group_id)) return true;
 
   // 语言库
   if (corpus(context)) return true;
@@ -176,14 +175,14 @@ https://github.com/Tsuk1ko/cq-picsearcher-bot`);
     return true;
   }
 
-  // setu
-  if (global.config.bot.setu.enable) {
-    if (sendSetu(context, logger)) return true;
-  }
-
   // reminder
   if (global.config.bot.reminder.enable) {
     if (rmdHandler(context)) return true;
+  }
+
+  // setu
+  if (global.config.bot.setu.enable) {
+    if (sendSetu(context)) return true;
   }
 
   //  反哔哩哔哩小程序
@@ -235,11 +234,11 @@ function adminPrivateMsg(e, context) {
   // Ban
   const { 'ban-u': bu, 'ban-g': bg } = args;
   if (bu && typeof bu === 'number') {
-    Logger.ban('u', bu);
+    logger.ban('u', bu);
     replyMsg(context, `已封禁用户${bu}`);
   }
   if (bg && typeof bg === 'number') {
-    Logger.ban('g', bg);
+    logger.ban('g', bg);
     replyMsg(context, `已封禁群组${bg}`);
   }
 
@@ -582,8 +581,6 @@ function doAkhr(context) {
     const imgs = getImgs(msg);
 
     const handleWords = words => {
-      // fix some ...
-      if (global.config.bot.akhr.ocr === 'ocr.space') words = _.map(words, w => w.replace(/冫口了/g, '治疗'));
       replyMsg(context, CQ.img64(Akhr.getResultImg(words)));
     };
 
@@ -718,6 +715,7 @@ function replySearchMsgs(context, ...msgs) {
     return asyncMap(msgs, msg =>
       bot('send_private_msg', {
         user_id: context.user_id,
+        group_id: global.config.bot.pmSearchResultTemp ? context.group_id : undefined,
         message: msg,
       })
     );