+++ src/cache.js
@@ -84,14 +84,13 @@ class PSCache {
    * @private
    */
   clearExpiredCache() {
+    const dir = Path.resolve(__dirname, '../data/pscache');
+    if (!Fse.existsSync(dir)) return;
     try {
-      klaw(Path.resolve(__dirname, '../data/pscache'), {
+      klaw(dir, {
         nodir: true,
         depthLimit: 1,
-        filter: ({ path, stats: { mtimeMs } }) => {
-          console.log(path, mtimeMs);
-          return mtimeMs < this.EXPIRE_MS;
-        },
+        filter: ({ stats: { mtimeMs } }) => mtimeMs < this.EXPIRE_MS,
       }).forEach(({ path }) => Fse.unlink(path));
     } catch (e) {
       console.error(`${global.getTime()} clear expired cache`);